requirements.txt
google-generativeai
feedparser
requests
praw
